FraudDetection/requirements.txt
numpy
scikit-learn
pandas
flask
matplotlib
pytest-cov


